// ticketmate_client/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type GetUserByGuidResponse = {
    user?: User;
};

export type User = {
    guid?: string;
    phoneNumber?: (string) | null;
    firstName?: (string) | null;
    lastName?: (string) | null;
    email?: (string) | null;
    avatar?: (string) | null;
    passwordHash?: (string) | null;
    isActive?: boolean;
};

export type PostProjectInsertProjectData = {
    query?: {
        Guid?: string;
        Name?: string;
    };
};

export type PostProjectInsertProjectResponse = (unknown);

export type PostProjectInsertProjectError = unknown;

export type GetProjectGetProjectByGuidData = {
    query?: {
        ProjectGuid?: string;
    };
};

export type GetProjectGetProjectByGuidResponse = (unknown);

export type GetProjectGetProjectByGuidError = unknown;

export type DeleteProjectDeleteProjectByGuidData = {
    query?: {
        Guid?: string;
    };
};

export type DeleteProjectDeleteProjectByGuidResponse = (unknown);

export type DeleteProjectDeleteProjectByGuidError = unknown;

export type PutProjectUpdateProjectByGuidData = {
    query?: {
        Guid?: string;
        IsActive?: boolean;
        Name?: string;
    };
};

export type PutProjectUpdateProjectByGuidResponse = (unknown);

export type PutProjectUpdateProjectByGuidError = unknown;

export type PostTicketInsertTicketData = {
    query?: {
        CreatedByUserGuid?: string;
        Description?: string;
        Guid?: string;
        PriorityId?: number;
        ProjectGuid?: string;
        StatusId?: number;
        Title?: string;
    };
};

export type PostTicketInsertTicketResponse = (unknown);

export type PostTicketInsertTicketError = unknown;

export type GetTicketGetTicketByGuidData = {
    query?: {
        TicketGuid?: string;
    };
};

export type GetTicketGetTicketByGuidResponse = (unknown);

export type GetTicketGetTicketByGuidError = unknown;

export type PutTicketUpdateTicketByGuidData = {
    query?: {
        Description?: string;
        Guid?: string;
        PriorityId?: number;
        StatusId?: number;
        Title?: string;
    };
};

export type PutTicketUpdateTicketByGuidResponse = (unknown);

export type PutTicketUpdateTicketByGuidError = unknown;

export type DeleteTicketDeleteTicketByGuidData = {
    query?: {
        TicketGuid?: string;
    };
};

export type DeleteTicketDeleteTicketByGuidResponse = (unknown);

export type DeleteTicketDeleteTicketByGuidError = unknown;

export type DeleteUserDeleteUserByGuidData = {
    query?: {
        UserGuid?: string;
    };
};

export type DeleteUserDeleteUserByGuidResponse = (unknown);

export type DeleteUserDeleteUserByGuidError = unknown;

export type GetUserGetUserByGuidData = {
    query?: {
        UserGuid?: string;
    };
};

export type GetUserGetUserByGuidResponse = (GetUserByGuidResponse);

export type GetUserGetUserByGuidError = unknown;

export type PostUserInsertUserData = {
    query?: {
        Avatar?: string;
        Email?: string;
        FirstName?: string;
        Guid?: string;
        IsActive?: boolean;
        LastName?: string;
        PasswordHash?: string;
        PhoneNumber?: string;
    };
};

export type PostUserInsertUserResponse = (unknown);

export type PostUserInsertUserError = unknown;

export type PutUserUpdateUserByGuidData = {
    query?: {
        Avatar?: string;
        Email?: string;
        FirstName?: string;
        Guid?: string;
        IsActive?: boolean;
        LastName?: string;
        PasswordHash?: string;
        PhoneNumber?: string;
    };
};

export type PutUserUpdateUserByGuidResponse = (unknown);

export type PutUserUpdateUserByGuidError = unknown;